# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Amazon Prime Video API – Spécification OpenAPI 3.0
  description: |
    Spécification complète simulant l'API d'Amazon Prime Video.

    ## Domaines couverts
    - **Auth** — Connexion, tokens JWT, 2FA
    - **Profils** — Gestion multi-profils (comme Netflix)
    - **Catalogue** — Films, séries, épisodes, documentaires
    - **Recherche** — Recherche textuelle, par genre, par acteur
    - **🔍 Recherche par image (IA)** — Identifier un film/acteur depuis une photo
    - **Lecteur** — Lecture, reprise, qualité vidéo, sous-titres
    - **Watchlist** — Liste à regarder plus tard
    - **Historique** — Contenus vus et reprises automatiques
    - **Évaluations & Avis** — Notes et commentaires
    - **Téléchargements** — Visionnage hors ligne
    - **Abonnement** — Prime Video, chaînes, add-ons
    - **Paiement** — Transactions, factures
    - **Recommandations** — Personnalisées, tendances, similaires
    - **Notifications** — Nouvelles sorties, rappels
    - **Contrôle parental** — Restrictions par profil
    - **📺 Publicités IA** — Génération automatique de bandes-annonces et spots publicitaires par IA
    - **Live & Events** — Streaming en direct, sport, événements
    - **Langues & Accessibilité** — Sous-titres, audiodescription, multi-langue

    ## Authentification
    Utiliser `POST /auth/login` pour obtenir un `accessToken` JWT (durée 1h).
    À inclure dans toutes les requêtes protégées : `Authorization: Bearer {accessToken}`

...
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via `POST /auth/login`. À inclure dans le header : `Authorization: Bearer {accessToken}`. Durée de validité : 3600 secondes (1 heure).
  schemas:
    Error:
      type: object
      description: |
        Réponse d'erreur standard. `code` est machine-readable (constante), `message` est lisible par un humain, `details` liste les erreurs de validation champ par champ.
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: Le contenu demandé est introuvable.
        details:
          type: array
          items:
            type: string
...
    Profile:
      type: object
      description: |
        Profil de visionnage sur un compte partagé (comme Netflix). Chaque profil a son propre historique, watchlist et recommandations. Un compte peut avoir jusqu'à 6 profils.
      properties:
        id:
          type: string
          example: prof_xyz789
        name:
          type: string
          example: Jean
        avatarUrl:
          type: string
          format: uri
          example: https://cdn.primevideo.com/avatars/profile_01.png
        isKidsProfile:
          type: boolean
          description: Profil enfant — restreint automatiquement aux contenus adaptés
        ageRating:
          type: string
          description: Classification d'âge maximale autorisée pour ce profil
          enum:
            - ALL
            - U
            - PG
            - PG_13
            - R_16
            - R_18
          example: R_18
        isPrimary:
          type: boolean
          description: Profil principal du compte (ne peut pas être supprimé)
        pin:
          type: string
          nullable: true
          description: Code PIN à 4 chiffres pour accéder à ce profil (optionnel)
        language:
          type: string
          example: fr
        createdAt:
          type: string
          format: date-time
    ProfileCreate:
      type: object
      description: Données pour créer un nouveau profil on le compte.
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
          example: Marie
        isKidsProfile:
          type: boolean
          default: false
        ageRating:
          type: string
          enum:
            - ALL
            - U
            - PG
            - PG_13
            - R_16
            - R_18
          default: R_18
        pin:
          type: string
          pattern: ^\d{4}$
          nullable: true
        language:
          type: string
          example: fr
        avatarUrl:
          type: string
          format: uri
    ProfileUpdate:
      type: object
      description: Champs modifiables d'un profil.
      properties:
        name:
          type: string
        isKidsProfile:
          type: boolean
        ageRating:
          type: string
          enum:
            - ALL
            - U
            - PG
            - PG_13
            - R_16
            - R_18
        pin:
          type: string
          pattern: ^\d{4}$
          nullable: true
        language:
          type: string
        avatarUrl:
          type: string
          format: uri
    ProfilePin:
      type: object
      description: Payload pour valider le PIN d'accès à un profil protégé.
      required:
        - pin
      properties:
        pin:
          type: string
          pattern: ^\d{4}$
          example: '1234'
...
  responses:
    Unauthorized:
      description: Non authentifié — token manquant, invalide ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token manquant ou invalide. Utilisez POST /auth/login pour obtenir un token.
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Erreur de validation — les données envoyées sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
...
    ProfileIdParam:
      name: profileId
      in: path
      required: true
      description: Identifiant unique du profil
      schema:
        type: string
        example: prof_xyz789
...
paths:
...
  /profiles:
    get:
      tags:
        - Profiles
      summary: Lister les profils du compte
      operationId: listProfiles
      description: |
        Liste tous les profils créés sur le compte. Chaque profil a son propre historique, watchlist, recommandations et paramètres. Un compte peut avoir jusqu'à 6 profils.
        C'est l'équivalent de la sélection de profil au démarrage de Netflix.
      responses:
        '200':
          description: Liste des profils
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profile'
                  maxProfiles:
                    type: integer
                    example: 6
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Profiles
      summary: Créer un profil
      operationId: createProfile
      description: |
        Crée un nouveau profil sur le compte. Les profils enfants (`isKidsProfile: true`) sont automatiquement restreints aux contenus adaptés (classification U et PG uniquement).
        Le nombre maximum de profils dépend du plan d'abonnement (généralement 6).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
            example:
              name: Marie
              isKidsProfile: false
              ageRating: R_18
              language: fr
      responses:
        '201':
          description: Profil créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Nombre maximum de profils atteint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: MAX_PROFILES_REACHED
                message: Vous avez atteint le nombre maximum de profils (6) pour votre abonnement.
  /profiles/{profileId}:
    parameters:
      - $ref: '#/components/parameters/ProfileIdParam'
    get:
      tags:
        - Profiles
      summary: Récupérer un profil
      operationId: getProfile
      description: Retourne le détail d'un profil spécifique avec ses paramètres.
      responses:
        '200':
          description: Profil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Profiles
      summary: Modifier un profil
      operationId: updateProfile
      description: |
        Modifie les paramètres d'un profil (nom, avatar, restrictions d'âge, PIN...). Seul le profil principal peut modifier les restrictions du compte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Profiles
      summary: Supprimer un profil
      operationId: deleteProfile
      description: |
        Supprime un profil et tout son historique, watchlist et données associées. Le profil principal (`isPrimary: true`) ne peut pas être supprimé. **Action irréversible.**
      responses:
        '204':
          description: Profil supprimé
        '400':
          description: Impossible de supprimer le profil principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /profiles/{profileId}/verify-pin:
    post:
      tags:
        - Profiles
      summary: Vérifier le PIN d'accès à un profil
      operationId: verifyProfilePin
      description: |
        Valide le code PIN d'un profil protégé. Retourne un token de session profil valable 4 heures. Ce token doit être inclus dans les requêtes suivantes via le header `X-Profile-Token`.
      parameters:
        - $ref: '#/components/parameters/ProfileIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePin'
      responses:
        '200':
          description: PIN validé — token de session profil retourné
          content:
            application/json:
              schema:
                type: object
                properties:
                  profileToken:
                    type: string
                    description: Token de session profil valable 4 heures
                  expiresIn:
                    type: integer
                    example: 14400
        '401':
          description: PIN incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...